feat (store) : 가게 단건 조회 구현 fix (store) : owner와 일반으로 파일 구분
