+++ src/main/java/com/example/outsourcing/domain/menu/repository/MenuRepository.java
@@ -4,7 +4,10 @@ import com.example.outsourcing.domain.menu.entity.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface MenuRepository extends JpaRepository<Menu, Long> {
 
+    List<Menu> findByStoreId(Long storeId);
 }
