+++ src/main/java/com/example/outsourcing/domain/store/repository/StoreRepository.java
@@ -4,6 +4,10 @@ import com.example.outsourcing.domain.store.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+
 public interface StoreRepository extends JpaRepository<Store, Long> {
 
     @Query("SELECT COUNT(s) " +
